new lexer baby
