+++ huff_core/tests/push_overrides.rs
@@ -15,7 +15,7 @@ fn test_gracefully_pads_push_override() {
 
     // Lex and Parse the source code
     let flattened_source = FullFileSource { source: OVERRIDEN_PUSH, file: None, spans: vec![] };
-    let lexer = lexer::LexerNew::new(flattened_source.source);
+    let lexer = Lexer::new(flattened_source.source);
     let tokens = lexer.into_iter().map(|x| x.unwrap()).collect::<Vec<Token>>();
     let mut parser = Parser::new(tokens, None);
 
@@ -52,7 +52,7 @@ fn test_constructs_exact_push_override() {
 
     // Lex and Parse the source code
     let flattened_source = FullFileSource { source: OVERRIDEN_PUSH, file: None, spans: vec![] };
-    let lexer = lexer::LexerNew::new(flattened_source.source);
+    let lexer = Lexer::new(flattened_source.source);
     let tokens = lexer.into_iter().map(|x| x.unwrap()).collect::<Vec<Token>>();
     let mut parser = Parser::new(tokens, None);
 
@@ -85,7 +85,7 @@ fn test_fails_on_push_underflow() {
     "#;
 
     let flattened_source = FullFileSource { source: OVERRIDEN_PUSH, file: None, spans: vec![] };
-    let lexer = lexer::LexerNew::new(flattened_source.source);
+    let lexer = Lexer::new(flattened_source.source);
     let tokens = lexer.into_iter().map(|x| x.unwrap()).collect::<Vec<Token>>();
     let mut parser = Parser::new(tokens, None);
     parser.parse().unwrap();
