+++ huff_lexer/README.md
@@ -24,44 +24,37 @@ use std::ops::Deref;
 // Instantiate a new lexer
 let source = "#define macro HELLO_WORLD()";
 let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-let mut lexer = Lexer::new(flattened_source);
+let mut lexer = Lexer::new(flattened_source.sour);
 
 // This token should be a Define identifier
 let tok = lexer.next().unwrap().unwrap();
-assert_eq!(tok, Token::new(TokenKind::Define, Span::new(0..7, None)));
-assert_eq!(lexer.current_span().deref(), &Span::new(0..7, None));
+assert_eq!(tok, Token::new(TokenKind::Define, Span::new(0..6, None)));
 
 // The next token should be the whitespace
 let tok = lexer.next().unwrap().unwrap();
-assert_eq!(tok, Token::new(TokenKind::Whitespace, Span::new(7..8, None)));
-assert_eq!(lexer.current_span().deref(), &Span::new(7..8, None));
+assert_eq!(tok, Token::new(TokenKind::Whitespace, Span::new(7..7, None)));
 
 // Then we should parse the macro keyword
 let tok = lexer.next().unwrap().unwrap();
-assert_eq!(tok, Token::new(TokenKind::Macro, Span::new(8..13, None)));
-assert_eq!(lexer.current_span().deref(), &Span::new(8..13, None));
+assert_eq!(tok, Token::new(TokenKind::Macro, Span::new(8..12, None)));
 
 // The next token should be another whitespace
 let tok = lexer.next().unwrap().unwrap();
-assert_eq!(tok, Token::new(TokenKind::Whitespace, Span::new(13..14, None)));
-assert_eq!(lexer.current_span().deref(), &Span::new(13..14, None));
+assert_eq!(tok, Token::new(TokenKind::Whitespace, Span::new(13..13, None)));
 
 // Then we should get the function name
 let tok = lexer.next().unwrap().unwrap();
-assert_eq!(tok, Token::new(TokenKind::Ident("HELLO_WORLD".to_string()), Span::new(14..25, None)));
-assert_eq!(lexer.current_span().deref(), &Span::new(14..25, None));
+assert_eq!(tok, Token::new(TokenKind::Ident("HELLO_WORLD".to_string()), Span::new(14..24, None)));
 
 // Then we should have an open paren
 let tok = lexer.next().unwrap().unwrap();
-assert_eq!(tok, Token::new(TokenKind::OpenParen, Span::new(25..26, None)));
-assert_eq!(lexer.current_span().deref(), &Span::new(25..26, None));
+assert_eq!(tok, Token::new(TokenKind::OpenParen, Span::new(25..25, None)));
 
 // Lastly, we should have a closing parenthesis
 let tok = lexer.next().unwrap().unwrap();
-assert_eq!(tok, Token::new(TokenKind::CloseParen, Span::new(26..27, None)));
-assert_eq!(lexer.current_span().deref(), &Span::new(26..27, None));
+assert_eq!(tok, Token::new(TokenKind::CloseParen, Span::new(26..26, None)));
 
+lexer.next();
 // We covered the whole source
-assert_eq!(lexer.current_span().end, source.len());
 assert!(lexer.eof);
 ```
