+++ huff_lexer/tests/comments.rs
@@ -15,7 +15,7 @@ use std::ops::Deref;
 fn instantiates() {
     let source = "#define macro HELLO_WORLD()";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let lexer = lexer::LexerNew::new(flattened_source.source.clone());
+    let lexer = Lexer::new(flattened_source.source.clone());
     assert!(!lexer.eof);
 }
 
@@ -23,7 +23,7 @@ fn instantiates() {
 fn single_line_comments() {
     let source = "// comment contents \n#define macro HELLO_WORLD()";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source.clone());
+    let mut lexer = Lexer::new(flattened_source.source.clone());
 
     // The first token should be a single line comment
     let tok = lexer.next();
@@ -98,7 +98,7 @@ fn single_line_comments() {
 fn multi_line_comments() {
     let source = "/* comment contents*/#define macro HELLO_WORLD()";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source.clone());
+    let mut lexer = Lexer::new(flattened_source.source.clone());
 
     // The first token should be a single line comment
     let tok = lexer.next();
