+++ huff_lexer/tests/decorators.rs
@@ -18,7 +18,7 @@ fn parses_decorator() {
         );
 
         let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-        let mut lexer = lexer::LexerNew::new(flattened_source.source.clone());
+        let mut lexer = Lexer::new(flattened_source.source.clone());
 
         let _ = lexer.next(); // whitespace
 
@@ -126,7 +126,7 @@ fn fails_to_parse_decorator_in_body() {
         );
 
         let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-        let mut lexer = lexer::LexerNew::new(flattened_source.source.clone());
+        let mut lexer = Lexer::new(flattened_source.source.clone());
 
         for token in lexer.by_ref() {
             if let Err(e) = token {
