+++ huff_lexer/tests/fsp.rs
@@ -7,7 +7,7 @@ use std::ops::Deref;
 fn free_storage_pointer() {
     let source = "FREE_STORAGE_POINTER() ";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source.clone());
+    let mut lexer = Lexer::new(flattened_source.source.clone());
 
     // The first token should be the fsp
     let tok = lexer.next().unwrap().unwrap();
