+++ huff_lexer/tests/keywords.rs
@@ -6,7 +6,7 @@ use std::ops::Deref;
 fn parses_macro_keyword() {
     let source = "#define macro";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source);
 
     // Define Identifier first
     let tok = lexer.next();
@@ -36,7 +36,7 @@ fn parses_macro_keyword() {
 fn parses_fn_keyword() {
     let source = "#define fn";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source);
 
     // Define Identifier first
     let tok = lexer.next();
@@ -66,7 +66,7 @@ fn parses_fn_keyword() {
 fn parses_test_keyword() {
     let source = "#define test";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source);
 
     // Define Identifier first
     let tok = lexer.next();
@@ -96,7 +96,7 @@ fn parses_test_keyword() {
 fn parses_function_keyword() {
     let source = "#define function";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source);
 
     // Define Identifier first
     let tok = lexer.next();
@@ -127,7 +127,7 @@ fn parses_function_keyword() {
 fn parses_event_keyword() {
     let source = "#define event TestEvent(uint256)";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source);
 
     // Define Identifier first
     let tok = lexer.next();
@@ -163,7 +163,7 @@ fn parses_event_keyword() {
 fn parses_constant_keyword() {
     let source = "#define constant";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source);
 
     // Define Identifier first
     let tok = lexer.next();
@@ -193,7 +193,7 @@ fn parses_constant_keyword() {
 fn parses_takes_and_returns_keywords() {
     let source = "#define macro TEST() = takes   (0)   returns   (0)";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source);
 
     let _ = lexer.next(); // #define
     let _ = lexer.next(); // whitespace
@@ -240,7 +240,7 @@ fn parses_takes_and_returns_keywords() {
 fn parses_takes_and_returns_keywords_tight_syntax() {
     let source = "#define macro TEST() = takes(0) returns(0)";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source);
 
     let _ = lexer.next(); // #define
     let _ = lexer.next(); // whitespace
@@ -285,7 +285,7 @@ fn parses_takes_and_returns_keywords_tight_syntax() {
 fn parses_function_type_keywords() {
     let source = "#define function test() view returns (uint256)";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source);
 
     let _ = lexer.next(); // #define
     let _ = lexer.next(); // whitespace
@@ -339,7 +339,7 @@ fn parses_function_definition_with_keyword_name() {
     for s in key_words {
         let source = &format!("#define function {s}(uint256) view returns(uint256)");
         let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-        let mut lexer = lexer::LexerNew::new(flattened_source.source);
+        let mut lexer = Lexer::new(flattened_source.source);
 
         let end_span_s = 17 + s.len();
 
@@ -411,14 +411,14 @@ fn parses_label_with_keyword_name() {
         );
 
         let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-        let mut lexer = lexer::LexerNew::new(flattened_source.source);
+        let mut lexer = Lexer::new(flattened_source.source);
 
         let tok = lexer.next();
         let unwrapped = tok.unwrap().unwrap();
         let fn_name_span = Span::new(0..s.len()-1, None);
         assert_eq!(unwrapped, Token::new(TokenKind::Label(s.to_string()), fn_name_span.clone()));
 
-       // let _ = lexer.next(); // colon
+        let _ = lexer.next(); // colon
         let _ = lexer.next(); // whitespace
 
         let tok = lexer.next();
@@ -461,7 +461,7 @@ fn parses_function_with_keyword_name() {
         let source = &format!("dup1 0x7c09063f eq {s} jumpi");
 
         let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-        let mut lexer = lexer::LexerNew::new(flattened_source.source);
+        let mut lexer = Lexer::new(flattened_source.source);
 
         let _ = lexer.next(); // dup1
         let _ = lexer.next(); // whitespace
@@ -518,7 +518,7 @@ fn parses_function_with_keyword_name_in_macro() {
         );
 
         let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-        let mut lexer = lexer::LexerNew::new(flattened_source.source);
+        let mut lexer = Lexer::new(flattened_source.source);
 
         let _ = lexer.next(); // whitespace
         let _ = lexer.next(); // #define
@@ -598,7 +598,7 @@ fn parses_keyword_arbitrary_whitespace() {
         let source = &format!("#define     {key}");
 
         let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-        let mut lexer = lexer::LexerNew::new(flattened_source.source);
+        let mut lexer = Lexer::new(flattened_source.source);
 
         // Define Identifier first
         let tok = lexer.next();
@@ -629,7 +629,7 @@ fn parses_keyword_arbitrary_whitespace() {
 fn parses_takes_keyword_arbitrary_whitespace() {
     let source = "#define macro TEST() =      takes (0) returns (0)";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source);
 
     let _ = lexer.next(); // #define
     let _ = lexer.next(); // whitespace
