+++ huff_lexer/tests/numbers.rs
@@ -6,7 +6,7 @@ use std::ops::Deref;
 fn lexes_zero_prefixed_numbers() {
     let source = "00";
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source);
 
     // The first and only token should be lexed as 0
     let tok = lexer.next().unwrap().unwrap();
@@ -22,7 +22,7 @@ fn lexes_zero_prefixed_numbers() {
 fn lexes_large_numbers() {
     let source = &format!("{}", usize::MAX);
     let flattened_source = FullFileSource { source, file: None, spans: vec![] };
-    let mut lexer = lexer::LexerNew::new(flattened_source.source);
+    let mut lexer = Lexer::new(flattened_source.source);
 
     // The first and only token should be lexed
     let tok = lexer.next().unwrap().unwrap();
